refactor: Rename Tasks to TaskList, update functionality to only interacting with list of tasks
